requirements.txt
asgiref==3.5.0
Django==4.0.4
sqlparse==0.4.2
tzdata==2022.1
django-filter~=21.1